requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0

# Experiment Tracking
wandb>=0.16.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Computer Vision
opencv-python>=4.8.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0
huggingface-hub>=0.19.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
